modified make_ptr so not a constructor/destructor, cleaned up makefiles, cleaned up calling to h2oaiglm_data(), ensured C++ gpu and cpu work on ipums.txt, ensured python gpu works on ipums.txt, and R works on simple.R.  Tried to make R work like C++ by removing weights, but wasn't successful for some reason -- it still gets error of order 0.25 on simple test, when should be 0.01 like C++
Former-commit-id: b1f6495bd2d6d8dfbff89d01188fd67fe357e69a
Former-commit-id: 944c1b8f1dea5b7409ede7351e20d4baf9afb833

+++ examples/cpp/reader.h
@@ -102,8 +102,8 @@ int fillData(size_t m, size_t n, // only used if name.empty()
     }
     cols--; //don't count target column
 
-    printf("rows: %d\n", rows); fflush(stdout);
-    printf("cols (w/o response): %d\n", cols); fflush(stdout);
+    printf("rows: %zu\n", rows); fflush(stdout);
+    printf("cols (w/o response): %zu\n", cols); fflush(stdout);
     ifs.close();
 
     size_t m = rows;
@@ -121,7 +121,7 @@ int fillData(size_t m, size_t n, // only used if name.empty()
       exit(1);
     } else {
       // Expect space-separated file with response in last column, no header
-      int maxlen = (n + 1) * 30;
+      size_t maxlen = (n + 1) * 30;
       char line[maxlen];
       const char sep[] = " ,";
